move `summary.go`, `summary_test.go` to `reporter` package from `scenarigo` package
